Configured for deployment to Heroku

+++ db/migrate/20120308015727_create_games.rb
@@ -1,15 +1,14 @@
 class CreateGames < ActiveRecord::Migration
   def change
     create_table :games do |t|
-      t.string :home_team
-      t.string :away_team
-      t.double :spread
-      t.double :over_under
-      t.string :game_time
-      t.double :home_score
-      t.double :away_score
-      t.string :game_id
-
+      t.string  :home_team
+      t.string  :away_team
+      t.decimal :spread
+      t.decimal :over_under
+      t.string  :game_time
+      t.decimal :home_score
+      t.decimal :away_score
+      t.string  :game_id
       t.timestamps
     end
   end
